requirements.txt
django-enumfield
